emulation: emulate counters only if counteren is not set
According to the RISC-V spec, [m|s]counteren determines whether or not
access to the corresponding hardware counter is permitted in the next
implemented privilege mode. When any bit is clear, attempts to read the
counter will cause an illegal-instruction exception.

riscv-pk sets all bits of counteren to ones during boot. It double-checks
the counteren during emulation of CSR instructions to ensure the
corresponding counter is disabled (not permitted) for the next privilege
mode.

The emulation aborts if the counteren is set, where the corresponding
attempt to read the counter should NOT cause the illegal-instruction
exception, and thus we are NOT expected to emulate it.