machine: emulate counters if scounteren is not writable
The implementation should make scounteren writable in order to let
M-mode software emulate counters. However, some implementations lack
such writable features currently and the ISA does not enforce this
requirement at least for now.

To provide a temp fix for such implementations, we emulate the counters
if scounteren is natively and totally unwritable.